test: isolate integration tests
