+++ .github/workflows/test.yml
@@ -9,7 +9,39 @@ on:
       - master
 
 jobs:
-  test:
+  unit:
+    strategy:
+      fail-fast: false
+      matrix:
+        go-version: [1.24.x]
+        platform: [ubuntu-latest]
+    runs-on: ${{ matrix.platform }}
+    steps:
+      - name: Install Go
+        uses: actions/setup-go@v5
+        with:
+          go-version: ${{ matrix.go-version }}
+
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Vet
+        run: go vet ./...
+
+      - name: Format check
+        run: |
+          unformatted=$(gofmt -l $(git ls-files '*.go'))
+          if [ -n "$unformatted" ]; then
+            echo "The following files are not formatted:" >&2
+            echo "$unformatted" >&2
+            exit 1
+          fi
+
+      - name: Unit tests
+        run: go test ./...
+
+  integration:
+    needs: unit
     strategy:
       fail-fast: false
       matrix:
@@ -34,17 +66,5 @@ jobs:
       - name: Confirm Docker
         run: docker info
 
-      - name: Vet
-        run: go vet ./...
-
-      - name: Format check
-        run: |
-          unformatted=$(gofmt -l $(git ls-files '*.go'))
-          if [ -n "$unformatted" ]; then
-            echo "The following files are not formatted:" >&2
-            echo "$unformatted" >&2
-            exit 1
-          fi
-
-      - name: Test
-        run: go test ./...
+      - name: Integration tests
+        run: go test -tags=integration ./...
