+++ core/runjob_integration_test.go
@@ -1,3 +1,6 @@
+//go:build integration
+// +build integration
+
 package core
 
 import (
